refactor: split disk exporter helpers
Move disk exporter lock, cleanup, audit file, and config helpers into focused package files so the main writer implementation stays easier to read.

Add shared test helpers for repeated disk writer setup in lifecycle regressions.
