+++ pkg/export/disk/disk.go
@@ -3,22 +3,13 @@ package disk
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"math"
 	"os"
-	"path"
-	"path/filepath"
-	"sort"
-	"strings"
 	"sync"
-	"syscall"
 	"time"
 
 	"github.com/open-policy-agent/gatekeeper/v3/pkg/export/util"
 	"github.com/open-policy-agent/gatekeeper/v3/pkg/logging"
-	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/client-go/util/retry"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
 
@@ -37,15 +28,6 @@ type Connection struct {
 	// current audit run file name
 	currentAuditRun string
 }
-
-// FailedConnection wraps a Connection with retry metadata.
-type FailedConnection struct {
-	Connection
-	FailedAt    time.Time
-	RetryCount  int
-	NextRetryAt time.Time
-}
-
 type Writer struct {
 	mu                           sync.Mutex
 	openConnections              map[string]Connection
@@ -81,142 +63,6 @@ var Connections = &Writer{
 }
 
 var log = logf.Log.WithName("disk-driver").WithValues(logging.Process, "export")
-
-func closeAndRemoveFilesWithRetry(conn *Connection) error {
-	return closeAndRemoveFilesWithBackoff(conn, retry.DefaultBackoff, os.RemoveAll)
-}
-
-func closeFileWithBackoff(conn *Connection, backoff wait.Backoff) error {
-	if conn.File == nil {
-		return nil
-	}
-	var lastErr error
-	err := wait.ExponentialBackoff(backoff, func() (bool, error) {
-		if err := conn.unlockAndCloseFile(); err != nil {
-			lastErr = fmt.Errorf("error closing file: %w", err)
-			return false, nil
-		}
-		conn.File = nil
-		lastErr = nil
-		return true, nil
-	})
-	if err != nil {
-		if lastErr != nil {
-			return lastErr
-		}
-		return err
-	}
-	return nil
-}
-
-func closeAndRemoveFilesWithBackoff(conn *Connection, backoff wait.Backoff, removeAll func(string) error) error {
-	var lastErr error
-	err := wait.ExponentialBackoff(backoff, func() (bool, error) {
-		if conn.File != nil {
-			if err := conn.unlockAndCloseFile(); err != nil {
-				lastErr = fmt.Errorf("error closing file: %w", err)
-				return false, nil
-			}
-			conn.File = nil
-		}
-		if err := removeAll(conn.Path); err != nil {
-			lastErr = fmt.Errorf("error deleting violations stored at old path: %w", err)
-			return false, nil
-		}
-		lastErr = nil
-		return true, nil
-	})
-	if err != nil {
-		if lastErr != nil {
-			return lastErr
-		}
-		return err
-	}
-	return nil
-}
-
-func (r *Writer) connectionLockLocked(connectionName string) *sync.Mutex {
-	if r.connectionLocks == nil {
-		r.connectionLocks = make(map[string]*sync.Mutex)
-	}
-	if r.connectionLocks[connectionName] == nil {
-		r.connectionLocks[connectionName] = &sync.Mutex{}
-	}
-	return r.connectionLocks[connectionName]
-}
-
-func (r *Writer) connectionLock(connectionName string, createIfMissing bool) (*sync.Mutex, bool) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	if !createIfMissing {
-		if _, exists := r.openConnections[connectionName]; !exists {
-			return nil, false
-		}
-	}
-	return r.connectionLockLocked(connectionName), true
-}
-
-// acquireCurrentConnectionLock returns the current per-connection lock held.
-// The caller must unlock the returned mutex.
-//
-// A connection lock can become stale if CloseConnection removes it after a
-// caller observes the lock but before the caller acquires it. Re-checking under
-// r.mu after acquiring the lock prevents callers from mutating connection state
-// while holding a lock that is no longer registered for connectionName.
-func (r *Writer) acquireCurrentConnectionLock(connectionName string, createIfMissing bool) (*sync.Mutex, bool) {
-	for {
-		connLock, exists := r.connectionLock(connectionName, createIfMissing)
-		if !exists {
-			return nil, false
-		}
-
-		connLock.Lock()
-
-		r.mu.Lock()
-		isCurrent := r.connectionLockIsCurrentLocked(connectionName, connLock)
-		r.mu.Unlock()
-		if isCurrent {
-			return connLock, true
-		}
-
-		connLock.Unlock()
-	}
-}
-
-func (r *Writer) connectionLockIsCurrentLocked(connectionName string, lock *sync.Mutex) bool {
-	return r.connectionLocks != nil && r.connectionLocks[connectionName] == lock
-}
-
-func (r *Writer) pathCleanupInProgressLocked(cleanupPath string) bool {
-	_, exists := r.cleanupPaths[cleanupPath]
-	return exists
-}
-
-func (r *Writer) pathInUseLocked(cleanupPath string) bool {
-	for _, conn := range r.openConnections {
-		if conn.Path == cleanupPath {
-			return true
-		}
-	}
-	return false
-}
-
-func (r *Writer) reserveCleanupPathLocked(cleanupPath string) bool {
-	if r.pathInUseLocked(cleanupPath) || r.pathCleanupInProgressLocked(cleanupPath) {
-		return false
-	}
-	if r.cleanupPaths == nil {
-		r.cleanupPaths = make(map[string]struct{})
-	}
-	r.cleanupPaths[cleanupPath] = struct{}{}
-	return true
-}
-
-func (r *Writer) releaseCleanupPath(cleanupPath string) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	delete(r.cleanupPaths, cleanupPath)
-}
 
 func (r *Writer) CreateConnection(_ context.Context, connectionName string, config interface{}) error {
 	path, maxResults, ttl, err := unmarshalConfig(config)
@@ -391,303 +237,4 @@ func (r *Writer) Publish(_ context.Context, connectionName string, data interfac
 	}
 	r.mu.Unlock()
 	return nil
-}
-
-func (conn *Connection) handleAuditStart(auditID string, topic string) error {
-	// Replace ':' with '_' to avoid issues with file names in windows
-	conn.currentAuditRun = strings.ReplaceAll(auditID, ":", "_")
-
-	dir := path.Join(conn.Path, topic)
-	if err := os.MkdirAll(dir, 0o777); err != nil {
-		return fmt.Errorf("failed to create directories: %w", err)
-	}
-
-	// Set the dir permissions to make sure reader can modify files if need be after the lock is released.
-	if err := os.Chmod(dir, 0o777); err != nil {
-		return fmt.Errorf("failed to set directory permissions: %w", err)
-	}
-
-	file, err := os.OpenFile(path.Join(dir, appendExtension(conn.currentAuditRun, "txt")), os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0o666)
-	if err != nil {
-		return fmt.Errorf("failed to open file: %w", err)
-	}
-	conn.File = file
-	err = retry.OnError(retry.DefaultBackoff, func(_ error) bool {
-		return true
-	}, func() error {
-		return syscall.Flock(int(conn.File.Fd()), syscall.LOCK_EX)
-	})
-	if err != nil {
-		return fmt.Errorf("failed to acquire lock: %w", err)
-	}
-	log.Info("Writing latest violations in", "filename", conn.File.Name())
-	return nil
-}
-
-func (conn *Connection) handleAuditEnd(topic string) error {
-	if err := retry.OnError(retry.DefaultBackoff, func(_ error) bool {
-		return true
-	}, conn.unlockAndCloseFile); err != nil {
-		return fmt.Errorf("error closing file: %w, %s", err, conn.currentAuditRun)
-	}
-	conn.File = nil
-
-	readyFilePath := path.Join(conn.Path, topic, appendExtension(conn.currentAuditRun, "log"))
-	if err := os.Rename(path.Join(conn.Path, topic, appendExtension(conn.currentAuditRun, "txt")), readyFilePath); err != nil {
-		return fmt.Errorf("failed to rename file: %w, %s", err, conn.currentAuditRun)
-	}
-	// Set the file permissions to make sure reader can modify files if need be after the lock is released.
-	if err := os.Chmod(readyFilePath, 0o777); err != nil {
-		return fmt.Errorf("failed to set file permissions: %w", err)
-	}
-	log.Info("File renamed", "filename", readyFilePath)
-
-	return conn.cleanupOldAuditFiles(topic)
-}
-
-func (conn *Connection) unlockAndCloseFile() error {
-	if conn.File == nil {
-		return fmt.Errorf("no file to close")
-	}
-	fd := int(conn.File.Fd())
-	if fd < 0 {
-		return fmt.Errorf("invalid file descriptor")
-	}
-	if err := syscall.Flock(fd, syscall.LOCK_UN); err != nil {
-		return fmt.Errorf("failed to release lock: %w", err)
-	}
-	if err := conn.File.Close(); err != nil {
-		return fmt.Errorf("failed to close file: %w", err)
-	}
-	return nil
-}
-
-func (conn *Connection) cleanupOldAuditFiles(topic string) error {
-	dirPath := path.Join(conn.Path, topic)
-	files, err := getFilesSortedByModTimeAsc(dirPath)
-	if err != nil {
-		return fmt.Errorf("failed removing older audit files, error getting files sorted by mod time: %w", err)
-	}
-	var errs []error
-	for i := 0; i < len(files)-conn.MaxAuditResults; i++ {
-		if e := os.Remove(files[i]); e != nil {
-			errs = append(errs, fmt.Errorf("error removing file: %w", e))
-		}
-	}
-
-	return errors.Join(errs...)
-}
-
-func getFilesSortedByModTimeAsc(dirPath string) ([]string, error) {
-	type fileInfo struct {
-		path    string
-		modTime time.Time
-	}
-	var filesInfo []fileInfo
-
-	err := filepath.Walk(dirPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		if !info.IsDir() {
-			filesInfo = append(filesInfo, fileInfo{path: path, modTime: info.ModTime()})
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	sort.Slice(filesInfo, func(i, j int) bool {
-		return filesInfo[i].modTime.Before(filesInfo[j].modTime)
-	})
-
-	var sortedFiles []string
-	for _, fi := range filesInfo {
-		sortedFiles = append(sortedFiles, fi.path)
-	}
-
-	return sortedFiles, nil
-}
-
-func appendExtension(name string, ext string) string {
-	return name + "." + ext
-}
-
-// validatePath checks if the provided path is valid and writable.
-func validatePath(path string) error {
-	if path == "" {
-		return fmt.Errorf("path cannot be empty")
-	}
-	if strings.Contains(path, "..") {
-		return fmt.Errorf("path must not contain '..', dir traversal is not allowed")
-	}
-	// validate if the path is writable
-	if err := os.MkdirAll(path, 0o777); err != nil {
-		return fmt.Errorf("failed to create directory: %w", err)
-	}
-	return nil
-}
-
-func unmarshalConfig(config interface{}) (string, float64, time.Duration, error) {
-	cfg, ok := config.(map[string]interface{})
-	if !ok {
-		return "", 0.0, 0, fmt.Errorf("invalid config format, expected map[string]interface{}")
-	}
-
-	path, pathOk := cfg[violationPath].(string)
-	if !pathOk {
-		return "", 0.0, 0, fmt.Errorf("missing or invalid 'path'")
-	}
-	if err := validatePath(path); err != nil {
-		return "", 0.0, 0, fmt.Errorf("invalid path: %w", err)
-	}
-	maxResults, maxResultsOk := cfg[maxAuditResults].(float64)
-	if !maxResultsOk {
-		return "", 0.0, 0, fmt.Errorf("missing or invalid 'maxAuditResults'")
-	}
-	if maxResults < 0 {
-		return "", 0.0, 0, fmt.Errorf("maxAuditResults cannot be negative")
-	}
-	if maxResults > maxAllowedAuditRuns {
-		return "", 0.0, 0, fmt.Errorf("maxAuditResults cannot be greater than the maximum allowed audit runs: %d", maxAllowedAuditRuns)
-	}
-	ttl := maxConnectionAge
-	if ttlStr, ok := cfg["closedConnectionTTL"].(string); ok {
-		duration, err := time.ParseDuration(ttlStr)
-		if err != nil {
-			return "", 0.0, 0, fmt.Errorf("invalid ttl format: %w", err)
-		}
-		ttl = duration
-	}
-	if ttl > maxConnectionAge {
-		return "", 0.0, 0, fmt.Errorf("closedConnectionTTL %s exceeds maximum allowed: %s", ttl, maxConnectionAge)
-	}
-	if ttl < minConnectionAge {
-		return "", 0.0, 0, fmt.Errorf("closedConnectionTTL %s is too short, must be at least 1 minute", ttl)
-	}
-	return path, maxResults, ttl, nil
-}
-
-// backgroundCleanup runs periodically to retry closing failed connections.
-// done is captured at goroutine launch time so the goroutine selects on an
-// immutable reference, avoiding a data race when CloseConnection reassigns
-// r.cleanupDone.
-func (r *Writer) backgroundCleanup(done <-chan struct{}) {
-	ticker := time.NewTicker(cleanupInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ticker.C:
-			r.retryFailedConnections()
-		case <-done:
-			log.Info("Background cleanup stopped")
-			return
-		}
-	}
-}
-
-// retryFailedConnections attempts to close connections that previously failed to close.
-func (r *Writer) retryFailedConnections() {
-	r.mu.Lock()
-
-	now := time.Now()
-	var toRemove []string
-	var toRetry []string
-
-	for name, failedConn := range r.closedConnections {
-		if now.Sub(failedConn.FailedAt) > failedConn.ClosedConnectionTTL {
-			log.Info("Removing expired failed connection", "connection", name, "age", now.Sub(failedConn.FailedAt))
-			toRemove = append(toRemove, name)
-			continue
-		}
-
-		if now.Before(failedConn.NextRetryAt) {
-			continue
-		}
-
-		if failedConn.RetryCount >= maxRetryAttempts {
-			log.Info("Max retry attempts exceeded for failed connection", "connection", name, "attempts", failedConn.RetryCount)
-			toRemove = append(toRemove, name)
-			continue
-		}
-
-		toRetry = append(toRetry, name)
-	}
-
-	for _, name := range toRemove {
-		delete(r.closedConnections, name)
-	}
-
-	// Collect connections that need retry and release the lock before I/O.
-	type retryItem struct {
-		name      string
-		conn      FailedConnection
-		closeOnly bool
-	}
-	items := make([]retryItem, 0, len(toRetry))
-	for _, name := range toRetry {
-		failedConn := r.closedConnections[name]
-		if r.pathInUseLocked(failedConn.Path) {
-			items = append(items, retryItem{name: name, conn: failedConn, closeOnly: true})
-			continue
-		}
-		if !r.reserveCleanupPathLocked(failedConn.Path) {
-			continue
-		}
-		items = append(items, retryItem{name: name, conn: failedConn})
-	}
-	r.mu.Unlock()
-
-	// Perform I/O outside the lock.
-	type retryResult struct {
-		name string
-		conn FailedConnection
-		ok   bool
-	}
-	results := make([]retryResult, 0, len(items))
-	for i := range items {
-		var err error
-		if items[i].closeOnly {
-			err = closeFileWithBackoff(&items[i].conn.Connection, retry.DefaultBackoff)
-		} else {
-			err = r.closeAndRemoveFilesWithRetry(&items[i].conn.Connection)
-			r.releaseCleanupPath(items[i].conn.Path)
-		}
-		if err == nil {
-			log.Info("Successfully closed previously failed connection", "connection", items[i].name)
-			results = append(results, retryResult{name: items[i].name, ok: true})
-		} else {
-			log.Info("Failed to close connection on retry", "connection", items[i].name, "error", err, "attempt", items[i].conn.RetryCount+1)
-			items[i].conn.RetryCount++
-			delay := time.Duration(float64(baseRetryDelay) * math.Pow(retryBackoffFactor, float64(items[i].conn.RetryCount)))
-			if maxRetryDelay > 0 && delay > maxRetryDelay {
-				delay = maxRetryDelay
-			}
-			delay = wait.Jitter(delay, Jitter)
-			items[i].conn.NextRetryAt = now.Add(delay)
-			results = append(results, retryResult{name: items[i].name, conn: items[i].conn})
-		}
-	}
-
-	// Re-acquire lock to update state.
-	r.mu.Lock()
-	defer r.mu.Unlock()
-
-	for i := range results {
-		if results[i].ok {
-			delete(r.closedConnections, results[i].name)
-		} else {
-			r.closedConnections[results[i].name] = results[i].conn
-		}
-	}
-
-	if len(r.closedConnections) > 0 {
-		log.Info("Failed connections remaining", "count", len(r.closedConnections))
-	} else {
-		log.Info("No failed connections remaining, cleanup done")
-		r.cleanupStopped = true
-		close(r.cleanupDone)
-	}
 }
